requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.27.0
boto3==1.34.51
Pillow==10.2.0
redis==5.0.2
twilio==8.13.0
python-dotenv==1.0.1
slowapi==0.1.9
aiofiles==23.2.1
